Added ability to control how the GlycanBuilder LogUtils class stores logger information.
Many of the exceptions that could be thrown by GlycanBuilder are instead stored as static properties of the LogUtils class.
Calling code is then expected to check these properties for any exceptions that may have been thrown.  Many of the
GlycanBuilder classes are used outside of a single-user environment, therefore users may see exceptions from other users.
This patch allows you to instruct LogUtils to set and get logger properties using an instance of LoggerStorageIndex. You
could create a version of this class which looks-up a thread-local copy of the logger properties (therefore isolating
multiple threads from each other).

(another example will be described shortly that is specific to Vaadin Navigator7 flows)